+++ test/acceptance/mapnik_meta_layergroup.js
@@ -55,6 +55,13 @@ describe('Create mapnik layergroup', function() {
         }
     };
 
+    function mapnikBasicLayerId(index) {
+        return 'layer' + index;
+    }
+    function typeLayerId(type, index) {
+        return type + '-' + mapnikBasicLayerId(index);
+    }
+
     it('with one mapnik layer should response with meta-stats for that layer', function(done) {
         var testClient = new TestClient({
             version: '1.4.0',
@@ -65,7 +72,8 @@ describe('Create mapnik layergroup', function() {
 
         testClient.createLayergroup(function(err, layergroup) {
             assert.ok(!err);
-            assert.ok(layergroup.metadata.layers[0].meta.stats[0].features === 5);
+            assert.equal(layergroup.metadata.layers[0].id, mapnikBasicLayerId(0));
+            assert.equal(layergroup.metadata.layers[0].meta.stats[0].features, 5);
             done();
         });
     });
@@ -81,8 +89,10 @@ describe('Create mapnik layergroup', function() {
 
         testClient.createLayergroup(function(err, layergroup) {
             assert.ok(!err);
-            assert.ok(layergroup.metadata.layers[0].meta.stats[0].features === 5);
-            assert.ok(layergroup.metadata.layers[1].meta.stats[0].features === 5);
+            assert.equal(layergroup.metadata.layers[0].id, mapnikBasicLayerId(0));
+            assert.equal(layergroup.metadata.layers[0].meta.stats[0].features, 5);
+            assert.equal(layergroup.metadata.layers[1].id, mapnikBasicLayerId(1));
+            assert.equal(layergroup.metadata.layers[1].meta.stats[0].features, 5);
             done();
         });
     });
@@ -99,9 +109,12 @@ describe('Create mapnik layergroup', function() {
 
         testClient.createLayergroup(function(err, layergroup) {
             assert.ok(!err);
-            assert.ok(layergroup.metadata.layers[0].meta.stats[0].features === 5);
-            assert.ok(layergroup.metadata.layers[1].meta.stats[0].features === 5);
-            assert.ok(layergroup.metadata.layers[2].meta.stats[0].features === 5);
+            assert.equal(layergroup.metadata.layers[0].id, mapnikBasicLayerId(0));
+            assert.equal(layergroup.metadata.layers[0].meta.stats[0].features, 5);
+            assert.equal(layergroup.metadata.layers[1].id, mapnikBasicLayerId(1));
+            assert.equal(layergroup.metadata.layers[1].meta.stats[0].features, 5);
+            assert.equal(layergroup.metadata.layers[2].id, mapnikBasicLayerId(2));
+            assert.equal(layergroup.metadata.layers[2].meta.stats[0].features, 5);
             done();
         });
     });
@@ -116,8 +129,9 @@ describe('Create mapnik layergroup', function() {
 
         testClient.createLayergroup(function(err, layergroup) {
             assert.ok(!err);
-            assert.ok(layergroup.metadata.layers[0].meta.stats[0].features === 5);
-            assert.ok(layergroup.metadata.layers[0].meta.stats[1].features === 5);
+            assert.equal(layergroup.metadata.layers[0].id, mapnikBasicLayerId(0));
+            assert.equal(layergroup.metadata.layers[0].meta.stats[0].features, 5);
+            assert.equal(layergroup.metadata.layers[0].meta.stats[1].features, 5);
             done();
         });
     });
@@ -133,10 +147,12 @@ describe('Create mapnik layergroup', function() {
 
         testClient.createLayergroup(function(err, layergroup) {
             assert.ok(!err);
-            assert.ok(layergroup.metadata.layers[0].meta.stats[0].features === 5);
-            assert.ok(layergroup.metadata.layers[0].meta.stats[1].features === 5);
-            assert.ok(layergroup.metadata.layers[1].meta.stats[0].features === 5);
-            assert.ok(layergroup.metadata.layers[1].meta.stats[1].features === 5);
+            assert.equal(layergroup.metadata.layers[0].id, mapnikBasicLayerId(0));
+            assert.equal(layergroup.metadata.layers[0].meta.stats[0].features, 5);
+            assert.equal(layergroup.metadata.layers[0].meta.stats[1].features, 5);
+            assert.equal(layergroup.metadata.layers[1].id, mapnikBasicLayerId(1));
+            assert.equal(layergroup.metadata.layers[1].meta.stats[0].features, 5);
+            assert.equal(layergroup.metadata.layers[1].meta.stats[1].features, 5);
             done();
         });
     });
@@ -152,10 +168,12 @@ describe('Create mapnik layergroup', function() {
 
         testClient.createLayergroup(function(err, layergroup) {
             assert.ok(!err);
-            assert.ok(layergroup.metadata.layers[0].meta.stats[0].features === 5);
+            assert.equal(layergroup.metadata.layers[0].id, mapnikBasicLayerId(0));
+            assert.equal(layergroup.metadata.layers[0].meta.stats[0].features, 5);
             assert.ok(!layergroup.metadata.layers[0].meta.stats[1]);
-            assert.ok(layergroup.metadata.layers[1].meta.stats[0].features === 5);
-            assert.ok(layergroup.metadata.layers[1].meta.stats[1].features === 5);
+            assert.equal(layergroup.metadata.layers[1].id, mapnikBasicLayerId(1));
+            assert.equal(layergroup.metadata.layers[1].meta.stats[0].features, 5);
+            assert.equal(layergroup.metadata.layers[1].meta.stats[1].features, 5);
             assert.ok(!layergroup.metadata.layers[1].meta.stats[2]);
             assert.ok(!layergroup.metadata.layers[2]);
             done();
@@ -173,8 +191,10 @@ describe('Create mapnik layergroup', function() {
 
         testClient.createLayergroup(function(err, layergroup) {
             assert.ok(!err);
+            assert.equal(layergroup.metadata.layers[0].id, mapnikBasicLayerId(0));
             assert.equal(layergroup.metadata.layers[0].type, 'mapnik');
-            assert.ok(layergroup.metadata.layers[0].meta.stats[0].features === 5);
+            assert.equal(layergroup.metadata.layers[0].meta.stats[0].features, 5);
+            assert.equal(layergroup.metadata.layers[1].id, typeLayerId('http', 0));
             assert.equal(layergroup.metadata.layers[1].type, 'http');
             done();
         });
@@ -191,8 +211,10 @@ describe('Create mapnik layergroup', function() {
 
         testClient.createLayergroup(function (err, layergroup) {
             assert.ok(!err);
+            assert.equal(layergroup.metadata.layers[0].id, typeLayerId('http', 0));
             assert.equal(layergroup.metadata.layers[0].type, 'http');
-            assert.ok(layergroup.metadata.layers[1].meta.stats[0].features === 5);
+            assert.equal(layergroup.metadata.layers[1].meta.stats[0].features, 5);
+            assert.equal(layergroup.metadata.layers[1].id, mapnikBasicLayerId(0));
             assert.equal(layergroup.metadata.layers[1].type, 'mapnik');
             done();
         });